requirements.txt
asgiref==3.4.1
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2021.5.30
cffi==1.14.6
channels==2.4.0
channels-redis==3.0.1
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
daphne==2.5.0
culqi==1.0.1
culqipy==0.2.6
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
